requirements.txt
graphviz==2.38
joblib==1.0.1
matplotlib==3.3.4
numpy==1.19.2
pandas==1.1.3
py-xgboost==0.90
python==3.6.8
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
shap==0.37.0
wheel==0.36.2